% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{estimateWeightedDispersions}
\alias{estimateWeightedDispersions}
\title{Estimate weights and weighted dispersions in edgeR analysis}
\usage{
estimateWeightedDispersions(d, design, maxit = 30, plot = FALSE,
  plotW = FALSE, verbose = TRUE, designZI = NULL, llTol = 1e-04,
  normalization = "TMM", normFactors = NULL, weights = NULL, ...)
}
\arguments{
\item{d}{DGEList object containing the gene expression counts.}

\item{design}{The design matrix specifying the experimental design.}

\item{maxit}{The number of iterations for the EM-algorithm. 30 by default but larger may be useful for large datasets (many samples). Convergence of the posterior probabilities can be checked by following the distribution of posterior probabilities over iterations with \code{plotW}.}

\item{plot}{Logical. Should the BCV plot be plotted in every iteration?}

\item{plotW}{Logical. Should the distribution of posterior probabilities for all zeros in the count matrix be plotted in every iteration?}

\item{...}{Other arguments passed on to a weighted version of \code{\link[edgeR]{estimateDisp}}}

\item{designZI.}{The design for the zero-excess model. If \code{NULL}, the effective library size is used by default.}
}
\description{
This function estimates posterior probabilities to belong to the count component in a zero-inflated negative binomial (ZINB) model and integrates these in dispersion estimation with \code{\link[edgeR]{estimateDisp}}.
}
\examples{
#' library(edgeR)
data(islamEset,package="zingeR")
islam=exprs(islamEset)[1:2000,]
design=model.matrix(~pData(islamEset)[,1])
d=DGEList(islam)
d=calcNormFactors(d)
d=estimateWeightedDispersions(d,design, maxit=200)
fit=glmFit(d,design)
lrt=glmWeightedF(fit,coef=2)
}
\seealso{
\code{\link[edgeR]{calcNormFactors}},  \code{\link[edgeR]{DGEList}}, \code{\link[edgeR]{estimateDisp}}
}

