% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getDatasetZTNB}
\alias{getDatasetZTNB}
\title{Estimate feature-wise parameters according to a ZTNB distribution.}
\usage{
getDatasetZTNB(counts, design, drop.extreme.dispersion = FALSE)
}
\arguments{
\item{counts}{the expression matrix}
}
\description{
Estimate feature-wise parameters according to a ZTNB distribution.
}

