% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTResults}
\alias{getTResults}
\title{Preliminary function for t-test using DESeq2.}
\usage{
getTResults(res, weights, filter)
}
\arguments{
\item{res}{The DESeq2 object obtained from \code{results}.}

\item{weights}{The zingeR weights}

\item{filter}{The variable to filter on.}
}
\description{
This function uses the test statistics in the standard DESeq2 object obtained from the \code{results} function to perform a t-test with degrees of freedom adjusted to zero-inflation downweighting.
}

