% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{islamEset}
\alias{islamEset}
\title{single-cell gene expression in mouse fibroblasts and embryonic stem cells (islam).}
\format{An object of class \code{ExpressionSet} with 11796 rows and 92 columns.}
\usage{
islamEset
}
\description{
The Islam dataset considers $44$ embryonic fibroblasts and $48$ embryonic stem cells in mouse. The original dataset was downloaded from the Gene Expression Omnibus (https://www.ncbi.nlm.nih.gov/geo) with accession number GSE29087. Control wells and genes with positive counts in fewer than 5 cells were removed.
}
\references{
Saiful Islam, Una Kja ̈llquist, Annalena Moliner, Pawel Zajac, Jian-Bing Fan, Peter L ̈onnerberg, and Sten Linnarsson. Characterization of the single-cell transcriptional landscape by highly multiplex RNA-seq. Genome research, 21(7):1160–7, jul 2011.
}
\keyword{datasets}

