% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lmWeightedFit}
\alias{lmWeightedFit}
\title{limma's lmFit but with ZI-adjusted residual df.}
\usage{
lmWeightedFit(object, design = NULL, ndups = 1, spacing = 1,
  block = NULL, correlation, weights = NULL, method = "ls", ...)
}
\arguments{
\item{object}{An \code{\link[limma]{EList-class}} object, usually produced by \code{\link[limma]{voom}} or \code{\link[zingeR]{weightedVoom}}.}

\item{design}{The design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be estimated.}

\item{weights}{A numeric matrix of observation weights.}

\item{...}{Other optional arguments to be passed to \code{lm.series}, \code{gls.series} or \code{mrlm}.}
}
\description{
This function is identical to the weighted linear model fit as provided by the \code{\link[limma]{lmFit}} function in the limma package, but adjust the model's residual degrees of freedom according to zero-inflation.
}
\details{
This function recycles the original \code{\link[limma]{lmFit}} function and adjusts the residual degrees of freedom according to zero-inflation as the sum of the posterior probabilities for a gene minus the number of coefficients to estimate in the model.
}
\examples{
library(edgeR)
data(islamEset,package="zingeR")
islam=exprs(islamEset)[1:2000,]
design=model.matrix(~pData(islamEset)[,1])
d=DGEList(islam)
nf=calcNormFactors(islam)
y=zeroWeightedVoom(d,design,nf=nf,maxit=200)
fit=lmWeightedFit(y,design)
fit=eBayes(fit)
#inference using e.g. topTable...
}
\references{
Law et al.: voom: precision weights unlock linear model analysis tools for RNA-seq read counts. Genome Biology 2014 15:R29
Smyth, G.K. (2004). Linear models and empirical Bayes methods for assessing di↵erential ex- pression in microarray experiments. Statistical Applications in Genetics and Molecular Biology 3, Article 3.
}
\seealso{
\code{\link[limma]{lmFit}}
}

