% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{zeroWeightedVoom}
\alias{zeroWeightedVoom}
\title{Estimate heteroscedastic and zero-inflation observation weights in limma-voom analysis}
\usage{
zeroWeightedVoom(d, design, nf, maxit = 30, plot = FALSE, plotW = FALSE,
  verbose = TRUE, designZI = NULL, llTol = 1e-04, ...)
}
\arguments{
\item{d}{DGEList object containing the gene expression counts.}

\item{design}{The design matrix specifying the experimental design.}

\item{maxit}{The number of iterations for the EM-algorithm. 30 by default but larger may be useful for large datasets (many samples). Convergence of the posterior probabilities can be checked by following the distribution of posterior probabilities over iterations with \code{plotW}.}

\item{plot}{Logical. Should the BCV plot be plotted in every iteration?}

\item{plotW}{Logical. Should the distribution of posterior probabilities for all zeros in the count matrix be plotted in every iteration?}

\item{designZI.}{The design for the zero-excess model. If \code{NULL}, the effective library size is used by default.}
}
\description{
This function estimates posterior probabilities to belong to the count component in a zero-inflated negative binomial (ZINB) model and integrates these with the heteroscedastic weights of the log transformed counts from limma-voom.
}
\examples{
library(limma)
data(islamEset,package="zingeR")
islam=exprs(islamEset)[1:2000,]
design=model.matrix(~pData(islamEset)[,1])
d=DGEList(islam)
nf=calcNormFactors(islam)
y=zeroWeightedVoom(d,design,nf=nf,maxit=200)
fit=lmWeightedFit(y,design)
fit=eBayes(fit)
tt=topTable(fit,coef=2,sort.by="none",number=nrow(fit))
}
\seealso{
\code{\link[edgeR]{calcNormFactors}},  \code{\link[edgeR]{DGEList}}, \code{\link[limma]{voom}}
}

