% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getDatasetZTNB}
\alias{getDatasetZTNB}
\title{Estimate feature-wise parameters according to a ZTNB distribution.}
\usage{
getDatasetZTNB(counts, design, drop.extreme.dispersion = FALSE,
  offset = NULL)
}
\arguments{
\item{counts}{A numeric matrix containing gene expression counts. Note that every gene in this matrix must have at least $p+1$ positive counts, with $p$ the number of columns in the design matrix.}

\item{design}{The design of the experiments with rows corresponding to samples and columns corresponding to coefficients.}

\item{drop.extreme.dispersion}{Either a numeric value between $0$ and $1$, stating the proportion of genes with extreme (high) dispersions to remove for simulation, or FALSE (default), if no dispersions should be removed for the analysis.}

\item{offset}{The offset to use (typically the sequencing depth) when estimating gene-wise means and dispersions in the zero-truncated negative binomial model. These parameters will be used as a basis for the simulation.}
}
\description{
This function fits a zero-truncated negative binomial distribution to the positive counts for each feature (row). By default, the zero-truncated negative binomial distribution from the \code{gamlss} package is used for this function.
}
\examples{
data(islamEset,package="zingeR")
islam=exprs(islamEset)[1:2000,]
design=model.matrix(~pData(islamEset)[,1])
params = getDatasetZTNB(counts=islam, design=design)
}
\seealso{
\code{\link[zingeR]{NBsimSingleCell}}
}

