% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{glmWeightedF}
\alias{glmWeightedF}
\title{MUST use F test. Chi-squared does not correct for ZI.}
\usage{
glmWeightedF(glmfit, coef = ncol(glmfit$design), contrast = NULL,
  test = "F", ZI = TRUE)
}
\description{
MUST use F test. Chi-squared does not correct for ZI.
}

