% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{zeroWeightsLS}
\alias{zeroWeightsLS}
\title{Estimate ZINB count component posterior probabilities}
\usage{
zeroWeightsLS(counts, design, maxit = 100, plot = FALSE, plotW = FALSE,
  verbose = TRUE, designZI = NULL, llTol = 1e-04)
}
\arguments{
\item{counts}{A count matrix with feature-wise expression values. Values in this matrix must be integers.}

\item{design}{Design matrix specifying the experimental design.}

\item{maxit}{The number of iterations for the EM-algorithm. 30 by default but larger may be useful for large datasets (many samples). Convergence of the posterior probabilities can be checked by following the distribution of posterior probabilities over iterations with \code{plotW}.}

\item{plot}{Logical. Should the BCV plot be plotted in every iteration?}

\item{plotW}{Logical. Should the distribution of posterior probabilities for all zeros in the count matrix be plotted in every iteration?}

\item{designZI.}{The design for the zero-excess model. If \code{NULL}, the effective library size is used by default.}
}
\description{
This function estimates posterior probabilities to belong to the count component according to a zero-inflated negative binomial (ZINB) model. Internally, edgeR is used for the estimation of the NB component.
}
\examples{
data(islamEset,package="zingeR")
islam=exprs(islamEset)[1:2000,]
design=model.matrix(~pData(islamEset)[,1])
zeroWeights=zeroWeightsLS(counts=islam, design=design, maxit=200)
}

